Patch non-existent import
This change patches a import which didn't exist. It addresses the
problem by using `self::uuid::Uuid` instead of `uuid::Uuid`. There
should be _no_ side effects. Honestly, I think this solution is
inelegant, and I'm not even sure how it fixes the issue.

I will look into it later, and try to find a more elegant solution to
the problem.

Signed-off-by: Dom Rodriguez <shymega@shymega.org.uk>
Tested-by: Dom Rodriguez <shymega@shymega.org.uk>