Add conditional to test for the "spawn" subcommand
This commit adds a conditional to test for the "spawn" subcommand for
Xtensis Core. Once the conditional matches, if at all, it prints to the
logger subsystem the initialisation string. After that, it
terminates.

This change was necessary as the subcommand "spawn" didn't have any
programming for it, and did nothing _at all_. Arguably, it still does do
nothing, but it does more than it did before. It has no side effects
other than interacting with the logging subsystem and printing a
string. It works as expected upon execution.

Acked-by: Dom Rodriguez <shymega@shymega.org.uk>
Reviewed-by: Dom Rodriguez <shymega@shymega.org.uk>
Tested-by: Dom Rodriguez <shymega@shymega.org.uk>
Signed-off-by: Dom Rodriguez <shymega@shymega.org.uk>