Refactor Xt and other changes
I have started the process of making Xt Core's intenal functionality a
modular design, which will help make component modification easier and
*more* flexible.

The major change here is of Xt's data structures (We have a Workspace
struct now!) and the moving of `xt-core/src` to `xt-core/rust/src`.

I have also moved build files, and added an initial logo
printing (on-startup of Xt Core) the Xt logo. This is a WIP feature, and
may not be released at this moment in time.

I have removed the Git tags, and I will release v0.1.0 once RPC is
implemented, and the modularisation is at a basic level of
functionality.

This is the first of many breaking changes to Xt Core, so here be
dragons!