remove: Xt logo printing code
This commit removes the code I added for printing an ASCII art
representation of the Xt logo on startup - although this was togglable
by a feature flag. Perhaps this could be added back in a later, more
mature version of Xt, but for the current state of Xt, it is really
_not_ necessary. This commit also removes the ASCII art logo as
well (rust/data/logo.txt).

This change is not anticipated to cause any severe breaking changes.